ФЗ 20

+++ telemetry/analyzer/src/main/java/ru/yandex/practicum/handler/hub/DeviceRemovedEventHandler.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.handler.hub;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.exception.NotFoundException;
 import ru.yandex.practicum.kafka.telemetry.event.DeviceRemovedEventAvro;
 import ru.yandex.practicum.kafka.telemetry.event.HubEventAvro;
 import ru.yandex.practicum.repository.SensorRepository;
@@ -23,6 +24,9 @@ public class DeviceRemovedEventHandler implements HubEventHandler {
     public void handle(HubEventAvro event) {
         DeviceRemovedEventAvro deviceRemovedEvent = (DeviceRemovedEventAvro) event.getPayload();
 
+        if (sensorRepository.findByIdAndHubId(deviceRemovedEvent.getId(), event.getHubId()).isEmpty()) {
+            throw new NotFoundException("Не найден сенсор с id = " + deviceRemovedEvent.getId() + " в пространстве hubId = " + event.getHubId());
+        }
         sensorRepository.deleteById(deviceRemovedEvent.getId());
         log.info("Из БД удален sensor/device с id  = {}", deviceRemovedEvent.getId());
     }
